第六节二维数组-50%

+++ java/src/bianliang/ZiMuZhuanHuan.java
@@ -12,7 +12,7 @@ public class ZiMuZhuanHuan {
         //需注意Scanner接收中没有单独接收char类型的函数，需要使用next()接收字符串，然后使用charAt(0)，取字符串的第一位
         char zimu = input.next().charAt(0);
         //a的ASCII的值为97，A的ASCII的值为65，各对应字母之前相差32
-        int result = (int)zimu;
+        int result = zimu;
         if(result >= 97){
             zimu = (char)(result - 32);
         }else {
